ProductDiscountDraft to interface. #942
