+++ sphere-models/src/main/java/io/sphere/sdk/productdiscounts/ProductDiscountDraft.java
@@ -1,55 +1,28 @@
 package io.sphere.sdk.productdiscounts;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import io.sphere.sdk.models.Base;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import io.sphere.sdk.models.LocalizedString;
 
 import javax.annotation.Nullable;
 
-public class ProductDiscountDraft extends Base {
-    private final ProductDiscountPredicate predicate;
-    private final LocalizedString name;
-    private final String sortOrder;
-    private final ProductDiscountValue value;
-    private final LocalizedString description;
-    private final Boolean active;
-
-    private ProductDiscountDraft(final LocalizedString name, final LocalizedString description, final ProductDiscountPredicate predicate, final ProductDiscountValue value, final String sortOrder, final boolean active) {
-        this.active = active;
-        this.predicate = predicate;
-        this.name = name;
-        this.sortOrder = sortOrder;
-        this.value = value;
-        this.description = description;
-    }
-
+@JsonDeserialize(as = ProductDiscountDraftImpl.class)
+public interface ProductDiscountDraft {
     @JsonProperty("isActive")
-    public Boolean isActive() {
-        return active;
-    }
+    Boolean isActive();
 
     @Nullable
-    public LocalizedString getDescription() {
-        return description;
-    }
+    LocalizedString getDescription();
 
-    public LocalizedString getName() {
-        return name;
-    }
+    LocalizedString getName();
 
-    public String getPredicate() {
-        return predicate.toSpherePredicate();
-    }
+    String getPredicate();
 
-    public String getSortOrder() {
-        return sortOrder;
-    }
+    String getSortOrder();
 
-    public ProductDiscountValue getValue() {
-        return value;
-    }
+    ProductDiscountValue getValue();
 
-    public static ProductDiscountDraft of(final LocalizedString name, final LocalizedString description, final ProductDiscountPredicate predicate, final ProductDiscountValue value, final String sortOrder, final boolean active) {
-        return new ProductDiscountDraft(name, description, predicate, value, sortOrder, active);
+    static ProductDiscountDraft of(final LocalizedString name, final LocalizedString description, final ProductDiscountPredicate predicate, final ProductDiscountValue value, final String sortOrder, final boolean active) {
+        return new ProductDiscountDraftImpl(name, description, predicate, value, sortOrder, active);
     }
 }
